+++ system/kissmvc_core.php
@@ -225,7 +225,7 @@ abstract class KISS_Engine
     public function requestNotFound($msg = '')
     {
         header("HTTP/1.0 404 Not Found");
-        die('<html><head><title>404 Not Found</title></head><body><h1>Not Found</h1><p>'.$msg.'<p>The requested URL was not found on this server.</p><p>Please go <a href="javascript: history.back( 1 )">back</a> and try again.</p><hr /><p>Powered By: <a href="http://kissmvc.com">KISSMVC</a></p></body></html>');
+        die('<html><head><title>404 Not Found</title></head><body><h1>Not Found</h1><p>'.$msg.'<p>The requested URL was not found on this server.</p><p>Please go <a href="javascript: history.back( 1 )">back</a> and try again.</p><hr /></body></html>');
     }
 }
 
@@ -543,7 +543,7 @@ abstract class KISS_Model
         if ($rs) {
             foreach ($rs as $key => $val) {
                 if (array_key_exists($key, $this->rs)) {
-                    $this->rs[$key] = $this->is_scalar_or_null($this->rs[$key]) ? $val : unserialize($this->COMPRESS_ARRAY ? gzinflate($val) : $val);
+                    $this->rs[$key] = $this->is_scalar_or_null($this->rs[$key]) ? $val : unserialize($this->COMPRESS_ARRAY ? gzinflate($val) : $val, ['allowed_classes' => false]);
                 }
             }
         }